# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Movie Search
  description: This is OpenAPI specification for the Movie Search Project
  version: 1.0.0
servers:
  - url: 'https://localhost/api/v1'
...
paths:
...
  /movies:
    get:
      tags:
        - Movies
      summary: Get existing movies.
      description: Get existing movies.
      operationId: getAllMovies
      parameters:
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/OrderParam'
        - $ref: '#/components/parameters/NameParam'
        - $ref: '#/components/parameters/GenreParam'
        - $ref: '#/components/parameters/CountryParam'
        - $ref: '#/components/parameters/ScoreFromParam'
        - $ref: '#/components/parameters/ScoreToParam'
        - $ref: '#/components/parameters/YearFromParam'
        - $ref: '#/components/parameters/YearToParam'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponseBase'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/MovieItem'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Movies
      summary: Add new movie.
      description: Add new movie.
      operationId: addNewMovie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieRequest'
      responses:
        '201':
          description: Successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: The same movie already exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /movies/{movieId}:
    get:
      tags:
        - Movies
      summary: Get movie by it's id.
      description: Get movie by it's id.
      operationId: getMovieById
      parameters:
        - name: movieId
          in: path
          description: ID of movie to be searching for
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '404':
          description: Movie with this id not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Movies
      summary: Update information about movie.
      description: Update information about movie.
      operationId: updateMovie
      parameters:
        - name: movieId
          in: path
          description: ID of movie to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieUpdateRequest'
      responses:
        '202':
          description: Successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '404':
          description: Movie with id not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Movie with this name already exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Country:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Germany
...
    Error:
      type: object
      properties:
        status:
          type: integer
          format: int64
          example: 404
        message:
          type: string
          example: Movie with this id not found
        timestamp:
          description: Time when error occurred in UTC±0:00
          type: string
          format: date-time
          example: "2025-06-11T10:15:30Z"
...
    CompanyItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 32
        name:
          type: string
          example: Warner Brothers
        logoUrl:
          type: string
          format: uri
          example: https://moviesearch.media/logos/paramount_pictures_logo.jpg
    Movie:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 5613
        name:
          type: string
          example: Fight Club
        slogan:
          type: string
          example: You do not talk about Fight Club
        description:
          type: string
          example: An insomniac office worker and a devil-may-care soap maker form an underground fight club that evolves into much more.
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
        countries:
          type: array
          items:
            $ref: '#/components/schemas/Country'
        companies:
          type: array
          items:
            $ref: '#/components/schemas/CompanyItem'
        score:
          type: number
          format: float
          minimum: 0.0
          maximum: 10.0
          example: 7.8
        coverUrl:
          type: string
          format: uri
          example: https://moviesearch.media/cover/fight_club.jpg
        trailerUrl:
          type: string
          format: uri
          example: https://youtu.be/BdJKm16Co6M?si=sY0RzPY-_H4wNSPE
        releaseDate:
          type: string
          format: date
          example: "2025-06-11"
        durationTime:
          description: Duration of movie in minutes.
          type: integer
          format: int32
          minimum: 0
          example: 127
        movieType:
          type: string
          default: MOVIE
          enum:
            - MOVIE
            - SERIES
          example: MOVIE
        movieRating:
          type: string
          default: UNRATED
          enum:
            - G
            - PG
            - PG-13
            - R
            - NC_17
            - UNRATED
          example: NC_17
        budget:
          type: integer
          format: int64
          example: 150000000
        revenue:
          type: integer
          format: int64
          example: 200000000
    MovieItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 213
        name:
          type: string
          example: The Shawshank Redemption
        coverUrl:
          type: string
          format: uri
          example: https://moviesearch.media/cover/shawshank_redemption.jpg
        releaseDate:
          type: string
          format: date
          example: "1994-06-11"
        score:
          type: number
          format: float
          minimum: 0.0
          maximum: 10.0
          example: 9.3
        durationTime:
          type: integer
          format: int32
          example: 146
        movieRating:
          type: string
          default: UNRATED
          enum:
            - G
            - PG
            - PG-13
            - R
            - NC_17
            - UNRATED
          example: R
    Genre:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 5
        name:
          type: string
          example: Horror
...
    MovieRequest:
      type: object
      properties:
        name:
          type: string
          example: Fight Club
        slogan:
          type: string
          example: You do not talk about Fight Club
        description:
          type: string
          example: An insomniac office worker and a devil-may-care soap maker form an underground fight club that evolves into much more.
        genreIds:
          type: array
          description: List of genre IDs the movie belongs to
          items:
            type: integer
            format: int64
          example: [ 1, 3, 7 ]
        countryIds:
          type: array
          description: List of country IDs related to the movie (e.g., production countries)
          items:
            type: integer
            format: int64
          example: [ 2, 5 ]
        companyIds:
          type: array
          description: List of companies IDs involved in the movie
          items:
            type: integer
            format: int64
          example: [ 4, 9 ]
        coverUrl:
          type: string
          format: uri
          example: https://moviesearch.media/cover/fight_club.jpg
        trailerUrl:
          type: string
          format: uri
          example: https://youtu.be/BdJKm16Co6M?si=sY0RzPY-_H4wNSPE
        releaseDate:
          type: string
          format: date
          example: "2025-06-11"
        durationTime:
          description: Duration of movie in minutes.
          type: integer
          format: int32
          minimum: 0
          example: 127
        movieType:
          type: string
          default: MOVIE
          enum:
            - MOVIE
            - SERIES
          example: MOVIE
        movieRating:
          type: string
          default: UNRATED
          enum:
            - G
            - PG
            - PG-13
            - R
            - NC_17
            - UNRATED
          example: NC_17
        budget:
          type: integer
          format: int64
          example: 150000000
        revenue:
          type: integer
          format: int64
          example: 200000000
    MovieUpdateRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 5613
        name:
          type: string
          example: Fight Club
        slogan:
          type: string
          example: You do not talk about Fight Club
        description:
          type: string
          example: An insomniac office worker and a devil-may-care soap maker form an underground fight club that evolves into much more.
        genreIds:
          type: array
          description: List of genre IDs the movie belongs to
          items:
            type: integer
            format: int64
          example: [ 1, 3, 7 ]
        countryIds:
          type: array
          description: List of country IDs related to the movie (e.g., production countries)
          items:
            type: integer
            format: int64
          example: [ 2, 5 ]
        companyIds:
          type: array
          description: List of companies IDs involved in the movie
          items:
            type: integer
            format: int64
          example: [ 4, 9 ]
        coverUrl:
          type: string
          format: uri
          example: https://moviesearch.media/cover/fight_club.jpg
        trailerUrl:
          type: string
          format: uri
          example: https://youtu.be/BdJKm16Co6M?si=sY0RzPY-_H4wNSPE
        releaseDate:
          type: string
          format: date
          example: "2025-06-11"
        durationTime:
          description: Duration of movie in minutes.
          type: integer
          format: int32
          minimum: 0
          example: 127
        movieType:
          type: string
          default: MOVIE
          enum:
            - MOVIE
            - SERIES
          example: MOVIE
        movieRating:
          type: string
          default: UNRATED
          enum:
            - G
            - PG
            - PG-13
            - R
            - NC_17
            - UNRATED
          example: NC_17
        budget:
          type: integer
          format: int64
          example: 150000000
        revenue:
          type: integer
          format: int64
          example: 200000000
...
    PaginatedResponseBase:
      type: object
      properties:
        totalItems:
          type: integer
          format: int64
          example: 142
        totalPages:
          type: integer
          format: int32
          example: 12
        currentPage:
          type: integer
          format: int32
          example: 2
        pageSize:
          type: integer
          format: int32
          example: 20
  parameters:
    SizeParam:
      name: size
      in: query
      description: Parameter for pagination. This parameter sets the size of response input.
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    #      explode: true
    PageParam:
      name: page
      in: query
      description: Parameter for pagination. This parameter sets the page of response input.
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SortParam:
      name: sort
      in: query
      description: Parameter for sorting. This parameter sets field to sort by.
      required: false
      schema:
        type: string
        example: name
    OrderParam:
      name: order
      in: query
      description: Parameter for sorting. This parameter sets the order to sort by.
      required: false
      schema:
        type: string
        default: asc
        enum:
          - asc
          - desc
    NameParam:
      name: name
      in: query
      description: Parameter for searching. This parameter sets the name to be searching by. For example (?name=Inception) it's going to search all movies or actors that have 'Inception' in name
      required: false
      schema:
        type: string
        example: Inception
    GenreParam:
      name: genreId
      in: query
      description: Parameter for filtering movies. This parameter sets the genres id to be searching by.
      required: false
      schema:
        type: array
        items:
          type: integer
        example: [1,5,9]
      style: form
      explode: true
    CountryParam:
      name: countryId
      in: query
      description: Parameter for filtering movies. This parameter sets the countries to be searching by.
      required: false
      schema:
        type: array
        items:
          type: integer
        uniqueItems: true
        example: [32,11]
      style: form
      explode: true
    ScoreFromParam:
      name: scoreFrom
      in: query
      description: Parameter for filtering movies. This parameter sets the min movie score to be searching by.
      required: false
      schema:
        type: number
        format: float
        minimum: 0.0
        maximum: 10.0
        example: 7.8
    ScoreToParam:
      name: scoreTo
      in: query
      description: Parameter for filtering movies. This parameter sets the max movie score to be searching by.
      required: false
      schema:
        type: number
        format: float
        minimum: 0.0
        maximum: 10.0
        example: 8.8
    YearFromParam:
      name: yearFrom
      in: query
      description: Parameter for filtering movies. This parameter sets the min year when movie was filmed to be searching by.
      required: false
      schema:
        type: integer
        minimum: 1900
        example: 1999
    YearToParam:
      name: yearTo
      in: query
      description: Parameter for filtering movies. This parameter sets the max year when movie was filmed to be searching by.
      required: false
      schema:
        type: integer
        minimum: 1900
        example: 1999